[util] move PRINCE cipher implementation to separate module
This moves the PRINCE cipher (Python) implementation to a separate
module so it can be reused between the ROM and flash image scrambling
scripts.
