+++ hw/ip/rom_ctrl/util/scramble_image.py
@@ -2,7 +2,6 @@
 # Copyright lowRISC contributors.
 # Licensed under the Apache License, Version 2.0, see LICENSE for details.
 # SPDX-License-Identifier: Apache-2.0
-
 '''Script for scrambling a ROM image'''
 
 import argparse
@@ -13,24 +12,12 @@ import hjson  # type: ignore
 from Crypto.Hash import cSHAKE256
 
 from mem import MemChunk, MemFile
-from util.design.secded_gen import ecc_encode_some, load_secded_config  # type: ignore
+from util.design.prince import prince, sbox  # type: ignore
+from util.design.secded_gen import ecc_encode_some  # type: ignore
+from util.design.secded_gen import load_secded_config
 
 ROM_BASE_WORD = 0x8000 // 4
 ROM_SIZE_WORDS = 8192
-
-PRINCE_SBOX4 = [
-    0xb, 0xf, 0x3, 0x2,
-    0xa, 0xc, 0x9, 0x1,
-    0x6, 0x7, 0x8, 0x0,
-    0xe, 0x5, 0xd, 0x4
-]
-
-PRINCE_SBOX4_INV = [
-    0xb, 0x7, 0x3, 0x2,
-    0xf, 0xd, 0x8, 0x9,
-    0xa, 0x6, 0x4, 0x0,
-    0x5, 0xe, 0xc, 0x1
-]
 
 PRESENT_SBOX4 = [
     0xc, 0x5, 0x6, 0xb,
@@ -46,54 +33,7 @@ PRESENT_SBOX4_INV = [
     0x0, 0x7, 0x9, 0xa
 ]
 
-PRINCE_SHIFT_ROWS64 = [
-    0x4, 0x9, 0xe, 0x3,
-    0x8, 0xd, 0x2, 0x7,
-    0xc, 0x1, 0x6, 0xb,
-    0x0, 0x5, 0xa, 0xf
-]
-
-PRINCE_SHIFT_ROWS64_INV = [
-    0xc, 0x9, 0x6, 0x3,
-    0x0, 0xd, 0xa, 0x7,
-    0x4, 0x1, 0xe, 0xb,
-    0x8, 0x5, 0x2, 0xf
-]
-
-PRINCE_ROUND_CONSTS = [
-    0x0000000000000000,
-    0x13198a2e03707344,
-    0xa4093822299f31d0,
-    0x082efa98ec4e6c89,
-    0x452821e638d01377,
-    0xbe5466cf34e90c6c,
-    0x7ef84f78fd955cb1,
-    0x85840851f1ac43aa,
-    0xc882d32f25323c54,
-    0x64a51195e0e3610d,
-    0xd3b5a399ca0c2399,
-    0xc0ac29b7c97c50dd
-]
-
-PRINCE_SHIFT_ROWS_CONSTS = [0x7bde, 0xbde7, 0xde7b, 0xe7bd]
-
 _UDict = Dict[object, object]
-
-
-def sbox(data: int, width: int, coeffs: List[int]) -> int:
-    assert 0 <= width
-    assert 0 <= data < (1 << width)
-
-    full_mask = (1 << width) - 1
-    sbox_mask = (1 << (4 * (width // 4))) - 1
-
-    ret = data & (full_mask & ~sbox_mask)
-    for i in range(width // 4):
-        nibble = (data >> (4 * i)) & 0xf
-        sb_nibble = coeffs[nibble]
-        ret |= sb_nibble << (4 * i)
-
-    return ret
 
 
 def subst_perm_enc(data: int, key: int, width: int, num_rounds: int) -> int:
@@ -166,113 +106,6 @@ def subst_perm_dec(data: int, key: int, width: int, num_rounds: int) -> int:
     return data ^ key
 
 
-def prince_nibble_red16(data: int) -> int:
-    assert 0 <= data < (1 << 16)
-    nib0 = (data >> 0) & 0xf
-    nib1 = (data >> 4) & 0xf
-    nib2 = (data >> 8) & 0xf
-    nib3 = (data >> 12) & 0xf
-    return nib0 ^ nib1 ^ nib2 ^ nib3
-
-
-def prince_mult_prime(data: int) -> int:
-    assert 0 <= data < (1 << 64)
-    ret = 0
-    for blk_idx in range(4):
-        data_hw = (data >> (16 * blk_idx)) & 0xffff
-        start_sr_idx = 0 if blk_idx in [0, 3] else 1
-        for nibble_idx in range(4):
-            sr_idx = (start_sr_idx + 3 - nibble_idx) % 4
-            sr_const = PRINCE_SHIFT_ROWS_CONSTS[sr_idx]
-            nibble = prince_nibble_red16(data_hw & sr_const)
-            ret |= nibble << (16 * blk_idx + 4 * nibble_idx)
-    return ret
-
-
-def prince_shiftrows(data: int, inv: bool) -> int:
-    assert 0 <= data < (1 << 64)
-    shifts = PRINCE_SHIFT_ROWS64_INV if inv else PRINCE_SHIFT_ROWS64
-
-    ret = 0
-    for nibble_idx in range(64 // 4):
-        src_nibble_idx = shifts[nibble_idx]
-        src_nibble = (data >> (4 * src_nibble_idx)) & 0xf
-        ret |= src_nibble << (4 * nibble_idx)
-    return ret
-
-
-def prince_fwd_round(rc: int, key: int, data: int) -> int:
-    assert 0 <= rc < (1 << 64)
-    assert 0 <= key < (1 << 64)
-    assert 0 <= data < (1 << 64)
-
-    data = sbox(data, 64, PRINCE_SBOX4)
-    data = prince_mult_prime(data)
-    data = prince_shiftrows(data, False)
-    data ^= rc
-    data ^= key
-    return data
-
-
-def prince_inv_round(rc: int, key: int, data: int) -> int:
-    assert 0 <= rc < (1 << 64)
-    assert 0 <= key < (1 << 64)
-    assert 0 <= data < (1 << 64)
-
-    data ^= key
-    data ^= rc
-    data = prince_shiftrows(data, True)
-    data = prince_mult_prime(data)
-    data = sbox(data, 64, PRINCE_SBOX4_INV)
-    return data
-
-
-def prince(data: int, key: int, num_rounds_half: int) -> int:
-    '''Run the PRINCE cipher
-
-    This uses the new keyschedule proposed by Dinur in "Cryptanalytic
-    Time-Memory-Data Tradeoffs for FX-Constructions with Applications to PRINCE
-    and PRIDE".
-
-    '''
-    assert 0 <= data < (1 << 64)
-    assert 0 <= key < (1 << 128)
-    assert 0 <= num_rounds_half <= 5
-
-    k1 = key & ((1 << 64) - 1)
-    k0 = key >> 64
-
-    k0_rot1 = ((k0 & 1) << 63) | (k0 >> 1)
-    k0_prime = k0_rot1 ^ (k0 >> 63)
-
-    data ^= k0
-    data ^= k1
-    data ^= PRINCE_ROUND_CONSTS[0]
-
-    for hri in range(num_rounds_half):
-        round_idx = 1 + hri
-        rc = PRINCE_ROUND_CONSTS[round_idx]
-        rk = k0 if round_idx & 1 else k1
-        data = prince_fwd_round(rc, rk, data)
-
-    data = sbox(data, 64, PRINCE_SBOX4)
-    data = prince_mult_prime(data)
-    data = sbox(data, 64, PRINCE_SBOX4_INV)
-
-    for hri in range(num_rounds_half):
-        round_idx = 11 - num_rounds_half + hri
-        rc = PRINCE_ROUND_CONSTS[round_idx]
-        rk = k1 if round_idx & 1 else k0
-        data = prince_inv_round(rc, rk, data)
-
-    data ^= PRINCE_ROUND_CONSTS[11]
-    data ^= k1
-
-    data ^= k0_prime
-
-    return data
-
-
 class Scrambler:
     subst_perm_rounds = 2
     num_rounds_half = 2
@@ -318,8 +151,7 @@ class Scrambler:
         return named_params
 
     @staticmethod
-    def _get_param_value(params: Dict[str, _UDict],
-                         name: str,
+    def _get_param_value(params: Dict[str, _UDict], name: str,
                          width: int) -> int:
         param = params.get(name)
         assert isinstance(param, dict)
@@ -378,7 +210,8 @@ class Scrambler:
         data_nonce_width = 64 - self._addr_width
         data_scr_nonce = self.nonce & ((1 << data_nonce_width) - 1)
         to_scramble = (data_scr_nonce << self._addr_width) | log_addr
-        full_keystream = prince(to_scramble, self.key, self.num_rounds_half)
+        full_keystream = int(
+            prince(to_scramble, self.key, self.num_rounds_half))
 
         return full_keystream & ((1 << width) - 1)
 
@@ -386,16 +219,16 @@ class Scrambler:
         assert self._addr_width < 64
         data_nonce_width = 64 - self._addr_width
         addr_scr_nonce = self.nonce >> data_nonce_width
-        return subst_perm_enc(log_addr, addr_scr_nonce,
-                              self._addr_width, self.subst_perm_rounds)
+        return subst_perm_enc(log_addr, addr_scr_nonce, self._addr_width,
+                              self.subst_perm_rounds)
 
     def addr_sp_dec(self, phy_addr: int) -> int:
         assert self._addr_width < 64
 
         data_nonce_width = 64 - self._addr_width
         addr_scr_nonce = self.nonce >> data_nonce_width
-        return subst_perm_dec(phy_addr, addr_scr_nonce,
-                              self._addr_width, self.subst_perm_rounds)
+        return subst_perm_dec(phy_addr, addr_scr_nonce, self._addr_width,
+                              self.subst_perm_rounds)
 
     def data_sp_enc(self, width: int, data: int) -> int:
         return subst_perm_enc(data, 0, width, self.subst_perm_rounds)
@@ -505,7 +338,8 @@ class Scrambler:
                 w39 = w32 | (chk_bits << 32)
                 clr39 = self.unscramble_word(39, log_addr, w39)
                 clr32 = clr39 & mask32
-                exp39 = ecc_encode_some(self.config, 'inv_hsiao', 32, [clr32])[0][0]
+                exp39 = ecc_encode_some(self.config, 'inv_hsiao', 32,
+                                        [clr32])[0][0]
                 if clr39 != exp39:
                     # The checksum doesn't match. Excellent!
                     found_mismatch = True
@@ -548,19 +382,19 @@ def main() -> int:
     # Check for collisions
     collisions = scr_mem.collisions()
     if collisions:
-        print('ERROR: This combination of ROM contents and scrambling\n'
-              '       key results in one or more collisions where\n'
-              '       different addresses have the same data.\n'
-              '\n'
-              '       Looks like we\'ve been (very) unlucky with the\n'
-              '       birthday problem. As a work-around, try again after\n'
-              '       generating some different RndCnst* parameters.\n',
-              file=sys.stderr)
+        print(
+            'ERROR: This combination of ROM contents and scrambling\n'
+            '       key results in one or more collisions where\n'
+            '       different addresses have the same data.\n'
+            '\n'
+            '       Looks like we\'ve been (very) unlucky with the\n'
+            '       birthday problem. As a work-around, try again after\n'
+            '       generating some different RndCnst* parameters.\n',
+            file=sys.stderr)
         print('{} colliding addresses:'.format(len(collisions)),
               file=sys.stderr)
         for addr0, addr1 in collisions:
-            print('  {:#010x}, {:#010x}'.format(addr0, addr1),
-                  file=sys.stderr)
+            print('  {:#010x}, {:#010x}'.format(addr0, addr1), file=sys.stderr)
         return 1
 
     scr_mem.write_vmem(args.outfile)
